infra: move indexing documentation test (#16595)
